Rename file for consistency purposes
